sikrer at repo er upto date

+++ backend/src/main/java/dashboard/controller/CompetitionController.java
@@ -11,27 +11,22 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/competitions")
 public class CompetitionController {
-
     @Autowired
     private CompetitionService competitionService;
-
     @GetMapping
     public List<CompetitionDTO> getAllCompetitions() {
         return competitionService.getAllCompetitions();
     }
-
     @GetMapping("/{competitionId}/athletes/accepted")
-    public List<AthleteDTO> getAcceptedAthletes(@PathVariable Long competitionId) {
+    public List<AthleteDTO> getAcceptedAthletes(@PathVariable int competitionId) {
         return competitionService.getAcceptedAthletes(competitionId);
     }
-
     @GetMapping("/{competitionId}/athletes/invited")
-    public List<AthleteDTO> getInvitedAthletes(@PathVariable Long competitionId) {
+    public List<AthleteDTO> getInvitedAthletes(@PathVariable int competitionId) {
         return competitionService.getInvitedAthletes(competitionId);
     }
-
     @GetMapping("/{competitionId}/athletes/checked-in")
-    public List<AthleteDTO> getCheckedInAthletes(@PathVariable Long competitionId) {
+    public List<AthleteDTO> getCheckedInAthletes(@PathVariable int competitionId) {
         return competitionService.getCheckedInAthletes(competitionId);
     }
 }
