+++ backend/src/main/java/dashboard/dto/CompetitionDTO.java
@@ -1,24 +1,24 @@
 package dashboard.dto;
 
 public class CompetitionDTO {
-    private Long id;
+    private int id;
     private String name;
     private String date;
 
     // Constructors, getters, and setters
     public CompetitionDTO() {}
 
-    public CompetitionDTO(Long id, String name, String date) {
+    public CompetitionDTO(int id, String name, String date) {
         this.id = id;
         this.name = name;
         this.date = date;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
