+++ backend/src/main/java/dashboard/repository/AthleteRepository.java
@@ -3,5 +3,5 @@ package dashboard.repository;
 import dashboard.model.Athlete;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AthleteRepository extends JpaRepository<Athlete, Long> {
+public interface AthleteRepository extends JpaRepository<Athlete, Integer> {
 }
