+++ backend/src/main/java/dashboard/service/CompetitionService.java
@@ -13,45 +13,38 @@ import java.util.stream.Collectors;
 
 @Service
 public class CompetitionService {
-
     @Autowired
     private CompetitionRepository competitionRepository;
-
     public List<CompetitionDTO> getAllCompetitions() {
         List<Competition> competitions = competitionRepository.findAll();
         return competitions.stream()
                 .map(this::convertToDTO)
                 .collect(Collectors.toList());
     }
-
-    public List<AthleteDTO> getAcceptedAthletes(Long competitionId) {
+    public List<AthleteDTO> getAcceptedAthletes(int competitionId) {
         Competition competition = competitionRepository.findById(competitionId).orElseThrow(() -> new RuntimeException("Competition not found"));
         return competition.getAthletes().stream()
                 .filter(Athlete::isAccepted)
                 .map(this::convertToAthleteDTO)
                 .collect(Collectors.toList());
     }
-
-    public List<AthleteDTO> getInvitedAthletes(Long competitionId) {
+    public List<AthleteDTO> getInvitedAthletes(int competitionId) {
         Competition competition = competitionRepository.findById(competitionId).orElseThrow(() -> new RuntimeException("Competition not found"));
         return competition.getAthletes().stream()
                 .filter(Athlete::isInvited)
                 .map(this::convertToAthleteDTO)
                 .collect(Collectors.toList());
     }
-
-    public List<AthleteDTO> getCheckedInAthletes(Long competitionId) {
+    public List<AthleteDTO> getCheckedInAthletes(int competitionId) {
         Competition competition = competitionRepository.findById(competitionId).orElseThrow(() -> new RuntimeException("Competition not found"));
         return competition.getAthletes().stream()
                 .filter(Athlete::isCheckedIn)
                 .map(this::convertToAthleteDTO)
                 .collect(Collectors.toList());
     }
-
     private CompetitionDTO convertToDTO(Competition competition) {
         return new CompetitionDTO(competition.getId(), competition.getName(), competition.getDate().toString());
     }
-
     private AthleteDTO convertToAthleteDTO(Athlete athlete) {
         return new AthleteDTO(
                 athlete.getId(),
